# api/openapi.yaml
openapi: 3.1.0
info:
  title: ISC Core API
  version: 0.1.0
  description: |
    ISC（接入编排器）内核的本地管理接口。

    ## 设计约束

    - **仅本机可达**：内核只监听命名管道 / Unix socket / 回环，绝不对外暴露。
      所有请求（含回环）都必须携带 `Authorization: Bearer <token>`，
      token 从数据目录下的 `runtime.json` 读取。见 docs/DECISIONS.md D09。
    - **快同步 + 慢异步**：列表与详情查询同步返回；所有可能耗时的操作
      （证书签发、凭据校验、防火墙应用、外部可达性验证）返回 `202` 与
      `job_id`，进度通过 `GET /v1/jobs/{id}` 或 WebSocket 事件流获取。
      见 docs/DECISIONS.md D08。
    - **契约优先**：本文件是接口的唯一真理。Go 服务端接口由 oapi-codegen 生成，
      前端类型由 openapi-typescript 生成。修改接口必须先改本文件。

    ## M0 阶段范围

    当前仅有元信息、健康检查、事件流与任务引擎。业务端点随里程碑加入。
servers:
  - url: http://localhost
    description: 实际地址由运行时的 runtime.json 决定（命名管道 / Unix socket / 回环）
...
security:
  - bearerAuth: []
paths:
...
  /v1/credentials:
    get:
      tags: [Credentials]
      operationId: listCredentials
      summary: 列出全部凭据
      description: |
        返回的凭据中，**敏感字段一律以掩码形式呈现**（例如 `abcd****wxyz`），
        明文永不离开内核。客户端提交更新时若某字段的值等于掩码，
        内核会保留原值而不是把掩码当作新值写入。
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: 凭据列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialList"
        default:
          $ref: "#/components/responses/Problem"

    post:
      tags: [Credentials]
      operationId: createCredential
      summary: 新建凭据
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialInput"
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credential"
        "400":
          $ref: "#/components/responses/Problem"
        "409":
          description: 同名凭据已存在
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/credentials/{id}:
    parameters:
      - $ref: "#/components/parameters/CredentialId"

    get:
      tags: [Credentials]
      operationId: getCredential
      summary: 凭据详情
      responses:
        "200":
          description: 凭据详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credential"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"

    patch:
      tags: [Credentials]
      operationId: updateCredential
      summary: 修改凭据
      description: |
        只更新请求体中出现的字段。敏感字段若提交掩码值则保留原值 ——
        这让"改个标签"不需要重新输入密钥。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialInput"
      responses:
        "200":
          description: 已更新
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credential"
        "400":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"

    delete:
      tags: [Credentials]
      operationId: deleteCredential
      summary: 删除凭据
      responses:
        "204":
          description: 已删除
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          description: 凭据仍被引用（例如有动态解析任务或证书在使用它）
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/Problem"

  # ---------------------------------------------------------------------------
  # 设置
  # ---------------------------------------------------------------------------
...
  /v1/config/export:
    get:
      tags: [Config]
      operationId: exportConfig
      summary: 导出配置
      description: |
        导出为 YAML，用于备份与迁移。

        `include_secrets` 默认为 **false**：导出的凭据字段为掩码，
        导入到另一台机器时需要重新填写密钥。这是刻意的默认值 ——
        一个"顺手导出"产生的明文密钥文件是常见的事故来源。
      parameters:
        - name: include_secrets
          in: query
          required: false
          description: 设为 true 才会导出明文凭据（务必妥善保管导出文件）。
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: 配置 YAML
          content:
            application/yaml:
              schema:
                type: string
        default:
          $ref: "#/components/responses/Problem"
  /v1/config/import:
    get:
      tags: [Config]
      operationId: describeImport
      summary: 说明导入行为
      description: |
        该端点只做说明，不做任何事。真正的导入用 POST。

        （保留它是因为 OpenAPI 3.1 的 `GET` 与 `POST` 语义需要分开表达，
        而客户端需要能发现"导入会做什么、会不会覆盖现有配置"。）
      responses:
        "200":
          description: 导入行为的说明
          content:
            application/json:
              schema:
                type: object
                properties:
                  dry_run_supported:
                    type: boolean
                  notes:
                    type: string
        default:
          $ref: "#/components/responses/Problem"

    post:
      tags: [Config]
      operationId: importConfig
      summary: 导入配置
      description: |
        `dry_run` 默认为 **true**：先返回"将会发生什么"（新建多少条、
        覆盖多少条、哪些会失败），确认后再以 `dry_run=false` 提交。

        这与 `docs/DECISIONS.md` D10 的"计划 → 预览 → 应用 → 回滚"
        是同一套思路：导入是破坏性操作，不该一步到位。
      parameters:
        - name: dry_run
          in: query
          required: false
          schema:
            type: boolean
            default: true
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              type: string
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: 导入结果（dry_run 时为预览）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        "400":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/config/import/ddns-go:
    post:
      tags: [Config]
      operationId: importDdnsGoConfig
      summary: 导入 ddns-go 配置
      description: |
        把 ddns-go 的 YAML 配置迁移为 ISC 的凭据与动态解析任务。

        迁移要点：
        - ddns-go 的每个 `dnsconf` 条目里，`dns.name` / `dns.id` / `dns.secret`
          被抽取为一个凭据（多条条目共用同一凭据时只建一个）；
        - `ipv4` / `ipv6` 的获取方式（网卡 / URL / 命令）与域名列表
          被转换为动态解析任务；
        - ddns-go 的 webhook 配置被转换为一个通知通道（M4 接入通知中心后生效）。
      parameters:
        - name: dry_run
          in: query
          required: false
          schema:
            type: boolean
            default: true
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              type: string
      responses:
        "200":
          description: 迁移结果（dry_run 时为预览）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        "400":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        token 位于数据目录的 `runtime.json` 中（字段 `token`），
        由内核首次启动时生成的 32 字节随机值经 base64url 编码得到。
  parameters:
...
    CredentialId:
      name: id
      in: path
      required: true
      description: 凭据 ID
      schema:
        type: string
        minLength: 1
    Cursor:
      name: cursor
      in: query
      required: false
      description: 分页游标，取自上一次响应的 next_cursor。
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      description: 单页条数。
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
    Problem:
      description: 错误响应（RFC 9457 problem+json）
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 定义的错误对象。
      required: [type, title, status]
      properties:
        type:
          type: string
          description: 标识错误类型的 URI。
          examples: ["https://isc.local/problems/not-found"]
        title:
          type: string
          description: 简短的人类可读摘要（已本地化）。
        status:
          type: integer
          description: HTTP 状态码。
          minimum: 100
          maximum: 599
        detail:
          type: string
          description: 本次错误的具体说明（已本地化）。
        instance:
          type: string
          description: 出错的具体请求路径。
        code:
          type: string
          description: |
            稳定的机器可读错误码，供客户端分支判断。
            与 title/detail 不同，该字段不随语言变化。
          examples: ["job_not_found", "job_not_cancelable"]
        request_id:
          type: string
          description: 请求 ID，便于在内核日志中定位。

    # ------------------------------------------------------------------
    # 元信息
    # ------------------------------------------------------------------
...
    ProviderCapabilities:
      type: object
      description: 该服务商实际支持的操作。GUI 据此置灰不支持的功能。
      required:
        - available
        - zone_list
        - record_list
        - record_create
        - record_update
        - record_delete
        - all_record_types
        - custom_ttl
        - proxy
        - dns01
      properties:
        available:
          type: boolean
          description: |
            该服务商的实现是否已就绪（目前指"能校验凭据"）。

            为 false 时，服务商出现在列表里只是为了让配置导入与界面展示
            完整，其能力位一律为 false。GUI 应当明确显示"尚未实现"，
            而不是让用户对着一排灰按钮猜原因。
        zone_list:
          type: boolean
          description: 能列出账号下的 DNS 区域。
        record_list:
          type: boolean
        record_create:
          type: boolean
        record_update:
          type: boolean
        record_delete:
          type: boolean
        all_record_types:
          type: boolean
          description: 支持 A/AAAA 之外的记录类型（CNAME/MX/TXT/NS/SRV/CAA 等）。
        custom_ttl:
          type: boolean
        proxy:
          type: boolean
          description: 支持 CDN 代理开关（如 Cloudflare 的橙云）。
        dns01:
          type: boolean
          description: 可用于 ACME DNS-01 证书校验。

    # ------------------------------------------------------------------
    # 凭据
    # ------------------------------------------------------------------
    Credential:
      type: object
      required: [id, provider, label, fields, created_at, updated_at]
      properties:
        id:
          type: string
        provider:
          type: string
        label:
          type: string
          description: 用户可读的备注名，用于在界面上区分同一服务商的多组凭据。
        fields:
          type: object
          additionalProperties:
            type: string
          description: |
            凭据字段。**敏感字段一律为掩码值**，明文永不通过接口返回。
            客户端提交更新时若原样回传掩码，内核会保留原值。
        capabilities:
          $ref: "#/components/schemas/ProviderCapabilities"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_verified_at:
          type: string
          format: date-time
        last_verify_ok:
          type: boolean
        last_verify_error:
          type: string
    CredentialInput:
      type: object
      required: [provider, label]
      properties:
        provider:
          type: string
        label:
          type: string
          minLength: 1
        fields:
          type: object
          additionalProperties:
            type: string
          description: |
            凭据字段明文。更新时只提交需要修改的字段；
            敏感字段若提交掩码值则保留原值。
    CredentialList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Credential"
        next_cursor:
          type: string

    # ------------------------------------------------------------------
    # 设置
    # ------------------------------------------------------------------
...
    ImportResult:
      type: object
      required: [dry_run, summary]
      properties:
        dry_run:
          type: boolean
        summary:
          $ref: "#/components/schemas/ImportSummary"
        warnings:
          type: array
          description: 不会导致失败、但用户应当知晓的问题（已本地化）。
          items:
            type: string
        errors:
          type: array
          description: 导致对应条目被跳过的错误（已本地化）。
          items:
            type: string
    ImportSummary:
      type: object
      properties:
        credentials_created:
          type: integer
        credentials_updated:
          type: integer
        credentials_skipped:
          type: integer
        tasks_created:
          type: integer
        tasks_updated:
          type: integer
        tasks_skipped:
          type: integer
        settings_updated:
          type: integer